News Summarizer Agent/backend/requirement.txt
transformers
torch
requests
python-dotenv
selenium
transformers
torch
webdriver-manager
fastapi
uvicorn